Big code reorganization.
