+++ flask_app/templates/blog/create.html
