+++ flask_app/templates/blog/index.html
