+++ setup.py
@@ -1,11 +1,16 @@
 from setuptools import find_packages, setup
 
 setup(
-    name="flaskr",
-    version="1.1.0",
+    name="tom",
+    version="0.0.1",
     packages=find_packages(),
     include_package_data=True,
     install_requires=[
         "flask",
+        "ortools",
+        "googlemaps",
+        "python-dotenv",
+        "pyyaml",
+        "wheel"
     ],
 )
