+++ tests/conftest.py
@@ -4,8 +4,8 @@ import yaml
 
 import pytest
 from dotenv import load_dotenv
-from flaskr import create_app
-from flaskr.db import get_db, init_db
+from flask_app import create_app
+from flask_app.db import get_db, init_db
 
 with open(os.path.join(os.path.dirname(__file__), "data.sql"), "rb") as f:
     _data_sql = f.read().decode("utf8")
