+++ tests/test_auth.py
@@ -1,6 +1,6 @@
 import pytest
 from flask import g, session
-from flaskr.db import get_db
+from flask_app.db import get_db
 
 
 def test_register(client, app):
