+++ tests/test_blog.py
@@ -1,5 +1,5 @@
 import pytest
-from flaskr.db import get_db
+from flask_app.db import get_db
 
 
 def test_index(client, auth):
