+++ tests/test_factory.py
@@ -1,4 +1,4 @@
-from flaskr import create_app
+from flask_app import create_app
 
 
 def test_config():
