+++ tests/test_trip.py
@@ -4,7 +4,7 @@ from tom.data_objects import Trip
 
 
 @pytest.mark.slow
-def test_optimize_trip(sample_trip: dict):
+def test_optimize_trip(env, sample_trip):
 
     trip = Trip.load_from_dict(sample_trip)
     solver = trip.optimize()
