--- tom/MANIFEST.in
@@ -1,4 +1,0 @@
-include flaskr/schema.sql
-graft flaskr/static
-graft flaskr/templates
-global-exclude *.pyc
