+++ tom/data_objects/trip.py
