--- tom/setup.cfg
@@ -1,7 +1,0 @@
-[tool:pytest]
-testpaths = tests
-
-[coverage:run]
-branch = True
-source = 
-    flaskr
